+++ CHANGELOG.md
@@ -12,6 +12,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Google authentication
 - Sign out method
 - Add main screen navigation
+- Login/password registration
+- Login/password authentication
 
 ### Changed